Added initial version bumping.
And simple tests thereof - certainly more is needed.
Currently all bumping lives in it's own method, and special care
is taken with dealing with the build and prerelease.

An alternative route could be to guard all the variables on behind
a property decorator, but to be honest, I think that is not that
pretty, except for the prerelease and build info - but that would
require more refactoring than I am comfortable doing in this
simple proof of concept.